LibWeb: Move selector parsing code to a separate file
